project/backend/requirements.txt
fastapi>=0.104.0
uvicorn>=0.24.0
python-jose[cryptography]>=3.3.0
passlib[bcrypt]>=1.7.4
python-multipart>=0.0.6
sqlalchemy>=2.0.23
aiosqlite>=0.19.0
python-dotenv>=1.0.0
pydantic>=2.4.2
pydantic-settings>=2.0.3
websockets>=12.0
python-jose[cryptography]>=3.3.0 